+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- feat: Add `throttle.d(delay, skipFirst)` sugar for dynamic function bodies
+- feat: Pass extra arguments into the function passed to `debounce.d()`
 - feat: Add `dbcdFn.cancel(true);` signature to finish (run) a debounced call
 - feat: Add `thrtldFn.finish(true);` signature to cancel the trailing call